// rw-app/src/utilities/interfaces/IVantagePro2PlusArchive.ts
import { DataStructureType } from "../DataStructureType";
import { SensorType } from "../SensorType";
import { ISensorData } from "./ISensorData";

export interface IVantagePro2PlusArchive extends ISensorData {
    temp_avg: number;
    temp_hi: number;
    temp_hi_at: number;
    TemperatureHighAt: Date;
    temp_lo: number;
    temp_lo_at: number;
    TemperatureLowAt: Date;
    hum_hi: number;
    hum_hi_at: number;
    HumidityHighAt: Date;
    hum_lo: number;
    hum_lo_at: number;
    HumidityLowAt: Date;
    dew_point_hi: number;
    dew_point_hi_at: number;
    DewPointHighAt: Date;
    dew_point_lo: number;
    dew_point_lo_at: number;
    DewPointLowAt: Date;
    wet_bulb_hi: number;
    wet_bulb_hi_at: number;
    WetBulbHighAt: Date;
    wet_bulb_lo: number;
    wet_bulb_lo_at: number;
    WetBulbLowAt: Date;
    wind_speed_avg: number;
    wind_speed_hi: number;
    wind_speed_hi_dir: number;
    wind_speed_hi_at: number;
    WindSpeedHighAt: Date;
    wind_chill_lo: number;
    wind_chill_lo_at: number;
    WindChillLowAt: Date;
    heat_index_hi: number;
    heat_index_hi_at: number;
    HeatIndexHighAt: Date;
    thw_index_hi: number;
    thw_index_hi_at: number;
    THWIndexHighAt: Date;
    thw_index_lo: number;
    thw_index_lo_at: number;
    THWIndexLowAt: Date;
    thsw_index_hi: number;
    thsw_index_hi_at: number;
    THSWIndexHighAt: Date;
    thsw_index_lo: number;
    thsw_index_lo_at: number;
    THSWIndexLowAt: Date;
    rainfall_clicks: number;
    rainfall_in: number;
    rainfall_mm: number;
    rain_rate_hi_in: number;
    rain_rate_hi_mm: number;
    rain_rate_hi_at: number;
    RainRateHighAt: Date;
    solar_rad_avg: number;
    solar_rad_hi: number;
    solar_rad_hi_at: number;
    SolarRadHighAt: Date;
    uv_index_avg: number;
    uv_index_hi: number;
    uv_index_hi_at: number;
    UVIndexHighAt: Date;
    wind_run: number;
    solar_energy: number;
    uv_dose: number;
    cooling_degree_days: number;
    heating_degree_days: number;
    arch_int: number;
    Type: SensorType;
    DataStructure: DataStructureType;
    Description: string;
}