// rw-app/src/utilities/interfaces/IAirLinkArchive.ts
import { DataStructureType } from "../DataStructureType";
import { SensorType } from "../SensorType";
import { ISensorData } from "./ISensorData";

export interface IAirLinkArchive extends ISensorData {
    temp_avg: number;
    temp_hi: number;
    temp_hi_at: number;
    TemperatureHighAt: Date;
    temp_lo: number;
    temp_lo_at: number;
    TemperatureLowAt: Date;
    hum_hi: number;
    hum_hi_at: number;
    HumidityHighAt: Date;
    hum_lo: number;
    hum_lo_at: number;
    HumidityLowAt: Date;
    dew_point_hi: number;
    dew_point_hi_at: number;
    DewPointHighAt: Date;
    dew_point_lo: number;
    dew_point_lo_at: number;
    DewPointLowAt: Date;
    wet_bulb_hi: number;
    wet_bulb_hi_at: number;
    WetBulbHighAt: Date;
    wet_bulb_lo: number;
    wet_bulb_lo_at: number;
    WetBulbLowAt: Date;
    heat_index_hi: number;
    heat_index_hi_at: number;
    HeatIndexHighAt: Date;
    pm_1_avg: number;
    pm_1_hi: number;
    pm_1_hi_at: number;
    pm_2p5_avg: number;
    pm_2p5_hi: number;
    pm_2p5_hi_at: number;
    PM2p5HighAt: Date;
    pm_10_avg: number;
    pm_10_hi: number;
    pm_10_hi_at: number;
    PM10HighAt: Date;
    pm_0p3_avg_num_part: number;
    pm_0p3_hi_num_part: number;
    pm_0p5_avg_num_part: number;
    pm_0p5_hi_num_part: number;
    pm_1_avg_num_part: number;
    pm_1_hi_num_part: number;
    pm_2p5_avg_num_part: number;
    pm_2p5_hi_num_part: number;
    pm_5_avg_num_part: number;
    pm_5_hi_num_part: number;
    pm_10_avg_num_part: number;
    pm_10_hi_num_part: number;
    aqi_type: string;
    aqi_avg_val: number;
    aqi_avg_desc: string;
    aqi_hi_val: number;
    aqi_hi_desc: string;
    arch_int: number;
    Type: SensorType;
    DataStructure: DataStructureType;
    Description: string;
}